move Finnhub API key server-side via /api/tickers route
